requirements.txt
SQLAlchemy==2.0.2
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0